module-3-gcp-big-query/python/requirements.txt
google-auth==2.3.0
google-auth-oauthlib==0.4.6
google-cloud-storage==1.41.1
pyarrow==15.0.0